They inctroduced a rumprun-bmk-ld. The linker no longer handles linking propperly. *shrugs*. So I fixed build system to compensate
