+++ apps/paws/makefile
@@ -1,12 +1,12 @@
-SRC = paws.c paws_mount_ext2.c
+SRC = paws.c paws_mount_ext2.c paws_test.c
 APP = paws
 LIBS = paws
 DEPS = backing.o
 
-main : backing.o
+main : tools/backing.o
 	@$(MAKE) all
 
-backing.o :
+tools/backing.o :
 	python ./tools/buildbacking.py
 
 include ../makefile.comp
