+++ apps/paws/paws_test.c
@@ -1,0 +1,167 @@
+#include "include/paws.h"
+
+static void
+printdirs(char dir[])
+{
+        DIR *dip = opendir(dir);
+        struct dirent *dp;
+
+        while((dp = readdir(dip)) != NULL) {
+                sleep(2);
+                printf("Current directory %s/%s\n", dir, dp->d_name);
+        }
+}
+
+static int
+testmkdir()
+{
+        int rv = -1;
+
+        if(mkdir("/mnt", 777) == 0){
+                printf("directory /mnt created\n");
+                rv = 0;
+        }
+
+        return rv;
+}
+
+/* mounting with ext2 mount type */
+static int
+testmount()
+{
+
+        struct ufs_args args;
+        struct stat sb;
+        int fd, rv;
+        size_t size;
+        char reader[200];
+        char readerfirst[100];
+        char writer[] = "abcdabcdabcabcdabcabcdabcdabcdabcabcdabcdabcdabcabcdabcabcdabcabcdabcabcdabcabcdabcddddddddd\n";
+        char writer2[] = "efgefgefgefgefgefefefefefgefgefgefgefefefefgefefggefefgefefefgefgefefgefgefgefgefggefgefggggggggggg\n";
+
+        fd = open("/dev/paws", O_RDWR);
+        assert(fd != -1);
+
+        rv = fstat(fd, &sb);
+        assert(rv == 0);
+
+        close(fd);
+
+        size = (size_t)sb.st_size;
+        args.fspec = "/dev/paws";
+        rv = mount_paws(MOUNT_EXT2FS, "/mnt", 0, &args, sizeof(args));
+        assert(rv != -1);
+
+        /* The backing image for paws already has some directories and a file that were placed in the host directory */
+        printf("\nfirst print:\n");
+        printdirs("/mnt");
+
+        rv = unmount_paws("/mnt", MNT_FORCE);
+        assert(rv != -1);
+
+        /* Mount 2nd time to write a few files... */
+        rv = mount_paws(MOUNT_EXT2FS, "/mnt", 0, &args, sizeof(args));
+        assert(rv != -1);
+
+        printf("\nMaking files...\n");
+
+        fd = open("/mnt/file", O_RDWR | O_APPEND | O_CREAT);
+        assert(fd > -1);
+        fd = open("/mnt/file1", O_RDWR | O_APPEND | O_CREAT);
+        assert(fd > -1);
+        fd = open("/mnt/file2", O_RDWR | O_APPEND | O_CREAT);
+        assert(fd > -1);
+
+        printf("\nSecond print: \n");
+        printdirs("/mnt");
+
+        rv = unmount_paws("/mnt", MNT_FORCE);
+        assert(rv != -1);
+
+        printf("\nUnmounted, reprinting /mnt\n");
+        printdirs("/mnt");
+
+        printf("\nRemounting and printing /mnt\n");
+        rv = mount_paws(MOUNT_EXT2FS, "/mnt", 0, &args, sizeof(args));
+        assert(rv != -1);
+        printf("\nThird print:\n");
+        printdirs("/mnt");
+
+        /* Writing to file 1 */
+        fd = open("/mnt/file1", O_RDWR | O_APPEND);
+        assert(fd > -1);
+        printf("Writting to file1: ");
+        rv = write(fd, writer, strlen(writer));
+        assert(rv > 0);
+        rv = write(fd, writer2, strlen(writer2));
+        assert(rv > 0);
+        close(fd);
+        printf("done\n");
+
+        rv = unmount_paws("/mnt", MNT_FORCE);
+        assert(rv != -1);
+
+        /* Remounting so we can read from file1 to check to see the changes are still there */
+        rv = mount_paws(MOUNT_EXT2FS, "/mnt", 0, &args, sizeof(args));
+        assert(rv != -1);
+
+        fd = open("/mnt/file1", O_RDWR | O_APPEND);
+        assert(fd > -1);
+
+        printf("\nReading from file 1:\n");
+        rv = read(fd, reader, strlen(writer)+strlen(writer2));
+        assert(rv > 0);
+
+        close(fd);
+
+        printf("%s\n", reader);
+
+        return 0;
+}
+
+static int
+testreadwrite()
+{
+        char reader[50];
+        char writer[] = "hello hobbes";
+        char writer2[] = " back for more\n";
+        int fd, rv;
+
+        fd = open("/dev/paws", O_RDWR);
+        assert(fd != -1);
+
+        printf("fd for /dev/paws: %d\n", fd);
+
+        rv = write(fd, writer, strlen(writer));
+        assert(rv > 0);
+
+        rv = write(fd, writer2, strlen(writer2));
+        assert(rv > 0);
+
+        lseek(fd, 0, SEEK_SET);
+
+        rv = read(fd, reader, strlen(writer)+strlen(writer2));
+        assert(rv > 0);
+
+        close(fd);
+
+        printf("%s\n", reader);
+
+        return 0;
+}
+
+int
+paws_tests()
+{
+        int rv;
+
+        /* testreadwrite(); */
+
+        rv = testmkdir();
+        assert(rv == 0);
+
+        rv = testmount();
+        assert(rv == 0);
+
+        return 0;
+}
